+++ Package/CHANGELOG.md
@@ -1,3 +1,38 @@
+## Version 0.1.8
+
+Features
+- Added various settings around scan nodes for items and enemies.
+
+Fixes
+- Fixed an issue where configuration would not generate if an enemy did not have an EnemyAI.
+
+## Version 0.1.7
+
+Features
+- Added group spawn count value for enemies.
+
+Fixes
+- Fixed the changelog.
+- Fixed an issue where trying to configure advanced dungeon properties would not work.
+
+## Version 0.1.6
+
+Fixes
+- Moved the 'warning' to the correct setting after some further issue diagnosing.
+
+## Version 0.1.5
+
+Features
+- Added more advanced dungeon configuration options.
+
+Fixes
+- Added a warning to configuring dungeon types due to reported desyncs (will fix whenever I find out the issue).
+
+## Version 0.1.4
+
+Fixes
+- Fixed dungeon configuration not recognizing the internal name of a dungeon flow.
+
 ## Version 0.1.3
 
 Features